backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
SQLAlchemy
alembic
psycopg[binary]
passlib[bcrypt]
bcrypt==3.2.2
python-jose[cryptography]
sentry-sdk
pytest
httpx
prometheus-client
pandas
pyarrow
redis>=5.0
rq>=2.0

